UPDATE: Add exports to components of Card
